+++ quartz-js/components/select/select.controller.js
@@ -8,7 +8,8 @@ vhxm.components.shared.select.controller = function(opts) {
 
     self.type = opts.type || 'standard';
     self.multiselect = opts.multiselect || false;
-    self.caret_position = opts.caret_position || 'right';
+    self.caret = opts.caret || 'right';
+    self.position = opts.position || 'bottom';
     self.model.items = opts.items;
 
     if (opts.selected) {
@@ -17,6 +18,9 @@ vhxm.components.shared.select.controller = function(opts) {
       });
     }
 
+    if (opts.isProcessing) {
+      self.state.isProcessing = opts.isProcessing;
+    }
     if (opts.onSelect) {
       self.state.onSelect = opts.onSelect;
     }
@@ -42,13 +46,11 @@ vhxm.components.shared.select.controller = function(opts) {
     let selected = opts.placeholder ? opts.placeholder : 'Select...';
 
     if (self.state.selected()) {
-      self.model.items().map(function(item) {
-        if (self.state.selected()[item[opts.prop_map.key]]) {
-          selected = self.state.selected()[item[opts.prop_map.key]].label;
-        }
-      });
       if (Object.keys(self.state.selected()).length > 1) {
         selected = 'Multiple items selected';
+      } else {
+        let key = Object.keys(self.state.selected())[0];
+        selected = self.state.selected()[key].label;
       }
     }
 
@@ -97,6 +99,11 @@ vhxm.components.shared.select.controller = function(opts) {
 
   self.selectItem = function(item, isInit) {
     let selected;
+    let obj = {
+      value: item[opts.prop_map.value],
+      label: item[opts.prop_map.label]
+    };
+
     if (!self.multiselect) {
       self.state.selected({});
     }
@@ -105,13 +112,9 @@ vhxm.components.shared.select.controller = function(opts) {
 
     if (selected[item[opts.prop_map.key]]) {
       delete selected[item[opts.prop_map.key]];
-      self.state.onSelect(null, self.state.selected(), 'removed');
+      self.state.onSelect(obj, self.state.selected(), 'removed');
     }
     else {
-      let obj = {
-        value: item[opts.prop_map.value],
-        label: item[opts.prop_map.label]
-      };
       selected[item[opts.prop_map.key]] = obj;
       self.state.onSelect(obj, self.state.selected(), 'added');
     }